separate core package
