--- src/export.ts
@@ -1,16 +1,0 @@
-export { isAdBlockInUse } from './lib/detect-ad-blocker';
-export { EventTimer } from './lib/event-timer';
-export { adSizes } from './lib/ad-sizes';
-export * as constants from './lib/constants';
-export {
-	bypassCommercialMetricsSampling,
-	initCommercialMetrics,
-} from './lib/send-commercial-metrics';
-export { buildPageTargeting } from './lib/targeting/build-page-targeting';
-export { postMessage } from './lib/messenger/post-message';
-export { buildImaAdTagUrl } from './lib/targeting/youtube-ima';
-export { getPermutivePFPSegments } from './lib/permutive';
-export type { AdsConfigDisabled } from './lib/types';
-export type { AdSize, SizeMapping, SlotName } from './lib/ad-sizes';
-export type { PageTargeting } from './lib/targeting/build-page-targeting';
-export type { AdsConfigUSNATorAus, AdsConfigTCFV2 } from './lib/types';
