+++ src/lib/ad-verification/prepare-ad-verification.ts
@@ -2,7 +2,7 @@ import { loadScript, log } from '@guardian/libs';
 import { refreshAdvert } from '../../display/load-advert';
 import { EventTimer } from '../../lib/event-timer';
 import { bypassCommercialMetricsSampling } from '../../lib/send-commercial-metrics';
-import type { ConfiantCallback } from '../../types/global';
+import type { ConfiantCallback } from '@guardian/commercial';
 import { getAdvertById } from '../dfp/get-advert-by-id';
 import { stripDfpAdPrefixFrom } from '../header-bidding/utils';
