--- tsconfig.core.json
@@ -1,13 +1,0 @@
-{
-	"compilerOptions": {
-		"baseUrl": "src",
-		"declaration": true,
-		"noEmit": false,
-		"outDir": "dist",
-		"typeRoots": ["./node_modules/@types", "./node_modules/web-vitals/dist"]
-	},
-	"extends": "./tsconfig.json",
-	"rootDir": "src",
-	"include": ["src/export.ts", "src/types"],
-	"exclude": ["**/*.spec.*", "**/*.test.*"]
-}
